+++ src/jhs/jhs.service.ts
@@ -34,6 +34,7 @@ export class JhsService {
       data: Prisma.JHStudentUpdateInput;
    }): Promise<JHStudent> {
       const { where, data } = params;
+      const { email, name , age } = data
       return this.prisma.jHStudent.update({
          where: where,
          data: data,
